Attach policies outside service account role
Currently the service-account-role module accepts an optional JSON
string. If present, it will be created as a policy and attached to the
created role. This provides a slight convenience with occasional blips
when applying terraform modules because the value is not known until
after apply, and Terraform gets confused by the for_each block that
makes the policy optional.

This moves the policies and attachments outside the module to eliminate
the dynamic portion of the module. It's slightly more verbose, but
clearer and eliminates the bugs during apply.