Reorganize operations platform
Rather than creating the workload platform as part of the operations
platform, assume the workload platform is applied first and then deploy
the operations platform on top.

This allows us to remove most of the variables and content for the
operations platform and inline the AWS workload values back into the
workload platform.