Disable annotations/labels in logs by default
Fluent Bit can add the pod annotations and labels to log output which
makes it possible to find and aggregate logs by labels. While this could
be useful, it creates extremely large log entries, which can be
expensive in Cloudwatch.

This removes most of the Kubernetes metadata from the log output. It
leaves namespace, container name, and pod name, which makes it possible
to filter by a particular application or process.