Add metrics for services
Every time series in Prometheus scraped from a service is tagged with a
"service" label[1]. This adds queries for metrics that are
pre-aggregated for a Kubernetes service and associates them with the
corresponding Kubernetes object.

This is useful for autoscaling metrics based on quantiles, which don't
make sense as a target for averageValue.

[1]: https://prometheus.io/docs/prometheus/latest/configuration/configuration/#service