Break up Prometheus instances
We currently have a single Prometheus instance per cluster, which
monitors cluster health, application health, and service health (ie
Istio). This Prometheus instance sends alerts to AlertManager and can
send metrics to an AWS Prometheus workspace for longer term metrics
storage and analysis. Prometheus Adapter talks to this instance for
pod autoscaling.

For a small cluster, this setup resulted in the following usage:

* 10% of one vCPU
* 600-700MiB RAM
* 50,000 metrics ingested per scrape

This results in heavy resource usage, expensive long-term storage, and
Prometheus instance that will require GiB of RAM at moderate scale.

In this change, the Prometheus instance is split into three:

* One instance to monitor cluster and pod health (kube-prometheus-stack)
* One instance to monitor Flightdeck and applications (flightdeck)
* One instance to collect aggregated metrics (federated)

The kube-prometheus-stack instance is used only for monitoring and
alerting about cluster and resource health.

The flightdeck instance is used for monitoring and alerting Flightdeck
and (optionally) applications, as well as pod autoscaling.

The federated instance scrapes aggregated metrics from other instances.
Metrics from this instance can be sent for long term storage.

For the same cluster, this setup requires the following usage:

* 5% of one vCPU, split up among three pods
* 600-700MiB RAM, split up among three pods
* 90 metrics ingested per scrape

This results in lower over all cluster resource usage, now distributed
to avoid a single point of failure and make it easier to distribute
loads as the cluster scales.

It sends vastly fewer metrics upstream for ingestion, resulting in lower
long-term storage costs.

It also provides a generic instance for federation, making it possible
to add other Prometheus instances which can then be aggregated and
ingested.