Add two prometheus replicas for high availability
flightdeck-prometheus is used for autoscaling as well, so if this goes
down for whatever reason or whatever node, then it would cause the site
to suffer performance problems. We can mitigate this risk by having
multiple nodes.

This also means currently our federated instance is getting twice as
much data, since it's pulling from both replicas and not being
deduplicated yet.