Remove SNS topic from network module
We currently create an SNS topic for alarms in the network module and
then attempt to find that topic based on tags or name in other modules.
This has always been an awkward place to create it, and in most
Flightdeck deployments we've actually used an existing topic.

This removes the topic that gets created in the network module. The only
place it's actually required in Flightdeck is for ingress. The ingress
module now accepts the same list of alarm actions that the underlying
ALB module requires.