Fix Fluent Bit after Helm chart upgrade
After upgrading the Helm chart version, we also need to update the
fluent bit version. This is because we're using the AWS fork of fluent
bit which is hosted on AWS and builds in support for AWS plugins.

Also adds a new test to verify that at a fluent bit pod is running on
each node.