Delete cluster before node role
Currently, when running `terraform destroy` for a cluster, it will
delete the IAM role used by EKS nodes before the cluster is fully
deleted. This results in elastic network interfaces and security groups
potentially being orphaned after the cluster is deleted as the cluster
loses permission to clean up the resources it created.

This uses `depends_on` to ensure that the IAM role is preserved until
the cluster is fully deleted.