Create log groups per Kubernetes namespace
Newer versions of Fluent Bit have a feature called "Record Accessor
Syntax" which allows you to interpolate labels from a log entry into the
filter settings. This allows us to make a log group based on the
Kubernetes namespace. You still need to have a default log group as a
fallback.

This contains the following changes:

- Retain cluster logs up to the retention period rather than deleting
  them with the cluster
- Create a log group based on the Kubernetes namespace
- Create a log stream based on the pod and container
- Allow Fluent Bit to create log groups in the given prefix