Add module for creating secret provider class
We currently add these to the manifests for applications, but that
involves copying and pasting names from the secrets we're creating.

This introduces a module that would manage the secretproviderclass in
Terraform, generating the necessary YAML to copy SecretsManager secrets
as environment variables.

This would mean the manifests only need to reference the
secretproviderclass in pod definitions rather than defining the secret
mappings in manifests.