Add AWS EKS IAM Auth Controller
We currently use the kubectl command to manually patch the aws-auth
configmap, as it's not possible to automatically adopt an existing
resource using Terraform. Users must also modify the platform
configuration any time they want to map a new IAM role to the cluster,
as there is only one place to do it.

The component EKS uses to map roles from the config map is the
[AWS IAM Authenticator for Kubernetes], which is maintained by AWS.
This component supports an IAMIdentityMapping CRD. This CRD would allow
merging entries into the configmap from multiple Kubernetes resources.
However, it's currently disabled on EKS.

There is a standalone [AWS EKS iam-auth-controller] maintained by the
community that will apply the AWS-designed CRD to the configmap in the
cluster.

This adds the AWS EKS IAM Auth Controller to replace the current
approach to patching the configmap, and uses it to automatically map
roles to groups where appropriate. This component can be removed if AWS
enables support for the CRD in EKS.

[AWS IAM Authenticator for Kubernetes]: https://github.com/kubernetes-sigs/aws-iam-authenticator
[AWS EKS iam-auth-controller]: https://github.com/rustrial/aws-eks-iam-auth-controller