Add AWS Managed Grafana module
Now that AWS Managed Grafana has an API for managing workspaces and the
Terraform AWS provider supports it, we can initialize our Grafana
workspaces using Terraform rather than doing it manually through the AWS
console.

This introduces a new module to create the workspace. It also updates
the API token, telemetry, and data source modules to take advantage of
the new workspace module.