Record Istio histogram buckets vs quantiles
We currently record quantiles for Istio histograms of response size and
request duration. This allows us to graph things like the 95th
percentile for request duration. However, it doesn't allow us to
calculate histograms that aren't prerecorded, and it doesn't allow us to
directly compare the bucket values for percentile-based SLOs.

This changes the Istio recording rules to record the aggregated bucket
values for request duration so that they'll be available for percentile
calculations in the managed Prometheus instance.