Convert jmesPath `object_alias` to `objectAlias`
The module expects jmesPath object` definitions with `object_alias` and
`path` keys. To be a valid jmesPath, SecretsManager requires converting
`object_alias` to its CamelCase equivalent.

Prior to this change no conversion was done which resulted in errors if
using `jmesPath`s. To maintain snakecase in the variable definition and
keep the documentation accurate, update the module to do the conversion.