Add custom group option to aws auth config map builder (#194)
At the moment, the custom_roles input is not suitable when we have to reuse a role for multiple groups.

This update will enable us to pass in the groups to be associated with a role using the custom_groups input instead.

the role will be the key, and the list of groups will be be value.